moved testable fontend code in another file
added frontend tests relates #58
