+++ public/index.html
@@ -30,7 +30,7 @@
   </section>
 </body>
 
+<script src="public/indextestable.js" type="text/javascript"></script>
 <script src="public/index.js" type="text/javascript"></script>
-
 
 </html>
