+++ tests/index.test.js
@@ -1,42 +1,25 @@
-// var test = require('tape');
-// var index = require('../public/index.js');
-//
-// console.log(index);
-//
-// // test urlBuilder function concatenates
-//
-// test('urlBuilder concatenates', function(t) {
-//   var actual = urlBuilder('apples');
-//   var expected = "http://localhost:4000/" + "search/" + 'apples';
-//   t.equal(actual, expected, 'should concatenate');
-//   t.end();
-// });
-//
-// // test retrive names from results object
-// var testObj = {
-//   results: [{
-//       name: "DarkRed",
-//       hex: "#8B0000"
-//     },
-//     {
-//       name: "IndianRed ",
-//       hex: "#CD5C5C"
-//     },
-//     {
-//       name: "MediumVioletRed",
-//       hex: "#C71585"
-//     },
-//     {
-//       name: "OrangeRed",
-//       hex: "#FF4500"
-//     },
-//     {
-//       name: "PaleVioletRed",
-//       hex: "#DB7093"
-//     },
-//     {
-//       name: "Red",
-//       hex: "#FF0000"
-//     }
-//   ]
-// }
+var test = require('tape');
+var index = require('../public/indextestable.js');
+
+//test cleanInput function
+test('cleanInput', function(t) {
+  var actual = index.cleanInput("apples1234_+/");
+  var expected = "apples";
+  t.equal(actual, expected, 'should return a string with all non alphabetical characters removed');
+  t.end();
+});
+test('cleanInput', function(t) {
+  var actual = index.cleanInput("aPPles");
+  var expected = "apples";
+  t.equal(actual, expected, "should return a string with all lower case characters");
+  t.end();
+});
+
+// test urlBuilder function concatenates
+
+test('urlBuilder', function(t) {
+  var actual = index.urlBuilder("apples");
+  var expected = "http://localhost:4000/" + "search/" + "apples";
+  t.equal(actual, expected, 'should concatenate');
+  t.end();
+});
